move the handwritten code into a derived class
