+++ python/EVPFFT-GUI/evpfft_gui/gui.py
@@ -2,13 +2,13 @@ import sys
 import os
 
 def main():
-    from evpfft_gui.ui_EVPFFT_GUI import Ui_MainWindow
+    from evpfft_gui.GUI import EVPFFT_GUI
     from PySide6.QtWidgets import QMainWindow, QApplication
     # MAIN WINDOW CLASS
     class MainWindow(QMainWindow):
         def __init__(self):
             QMainWindow.__init__(self)
-            self.ui = Ui_MainWindow()
+            self.ui = EVPFFT_GUI()
             self.ui.setupUi(self)
             
             # SHOW WINDOW
