+++ python/EVPFFT-GUI/evpfft_gui/ui_EVPFFT_GUI.py
@@ -1273,22 +1273,6 @@ class Ui_MainWindow(object):
         self.verticalLayout_20.setContentsMargins(0, 0, 0, 0)
         self.Paraview = QWidget(self.ParaviewWindow)
         self.Paraview.setObjectName(u"Paraview")
-        
-        # ================== New imports =====================
-        # Paraview imports
-        import paraview.simple as pvsimple
-        from vtkmodules.qt.QVTKRenderWindowInteractor import QVTKRenderWindowInteractor
-        render_view = pvsimple.CreateRenderView()
-        self.Paraview = QVTKRenderWindowInteractor(rw=render_view.GetRenderWindow(),iren=render_view.GetInteractor())
-        
-        # Open Url in Help Menu
-        from PySide6.QtCore import QUrl
-        from PySide6.QtGui import QDesktopServices
-        def openUrl():
-            url = QUrl('https://lanl.github.io/Fierro/')
-            if not QDesktopServices.openUrl(url):
-                QMessageBox.warning(self, 'Open Url', 'Could not open url')
-        # ======================================================
 
         self.verticalLayout_20.addWidget(self.Paraview)
 
@@ -1364,801 +1348,6 @@ class Ui_MainWindow(object):
 
 
         QMetaObject.connectSlotsByName(MainWindow)
-        
-        # ===================== New imports =======================
-        import os, sys
-        import numpy as np
-        from PySide6.QtWidgets import (QFileDialog, QMessageBox)
-        from PySide6.QtCore import (QTimer, QProcess)
-        import matplotlib
-        matplotlib.use('qt5agg')
-        from matplotlib.backends.backend_qtagg import FigureCanvasQTAgg
-        from matplotlib.figure import Figure
-        from matplotlib.backends.backend_qtagg import NavigationToolbar2QT
-        import csv
-        import re
-        import shutil
-        
-        # BUTTON SETUP
-        # Connect tab buttons to settings windows
-        self.BImportPart.clicked.connect(lambda: self.ToolSettings.setCurrentIndex(1))
-        self.BDefineMaterial.clicked.connect(lambda: self.ToolSettings.setCurrentIndex(2))
-        self.BApplyBC.clicked.connect(lambda: self.ToolSettings.setCurrentIndex(3))
-        self.BSolverSettings.clicked.connect(lambda: self.ToolSettings.setCurrentIndex(4))
-        self.BViewResults.clicked.connect(lambda: self.ToolSettings.setCurrentIndex(5))
-        
-        # Help menu
-        self.actionEVPFFT_Manual.triggered.connect(openUrl)
-        
-        # Upload Geometry
-        def geometry_upload_click():
-            try:
-                self.stl
-            except:
-                print('')
-            else:
-                pvsimple.Delete(self.stl)
-                
-            try:
-                self.voxel_reader
-            except:
-                print('')
-            else:
-                pvsimple.Delete(self.threshold)
-                
-            global b3_filename
-            b3_filename = QFileDialog.getOpenFileName(
-                filter="Geometry File (*.stl *.vtk)",
-            )
-            # Paraview window
-            self.file_type = b3_filename[0][-4:-1]
-            if self.file_type == '.st':
-                self.stl = pvsimple.STLReader(FileNames = b3_filename)
-                self.STLVoxelization.setEnabled(True)
-                self.LNumberOfVoxelsX.setEnabled(True)
-                self.INNumberOfVoxelsX.setEnabled(True)
-                self.LNumberOfVoxelsY.setEnabled(True)
-                self.INNumberOfVoxelsY.setEnabled(True)
-                self.LNumberOfVoxelsZ.setEnabled(True)
-                self.INNumberOfVoxelsZ.setEnabled(True)
-                self.BVoxelizeGeometry.setEnabled(True)
-            elif self.file_type == '.vt':
-                self.stl = pvsimple.LegacyVTKReader(FileNames = b3_filename)
-                pvsimple.SetDisplayProperties(Representation = "Surface")
-                shutil.copy(b3_filename[0], VTK_OUTPUT)
-                self.INNumberOfVoxelsX.setText(QCoreApplication.translate("MainWindow", u"32", None))
-                self.INNumberOfVoxelsY.setText(QCoreApplication.translate("MainWindow", u"32", None))
-                self.INNumberOfVoxelsZ.setText(QCoreApplication.translate("MainWindow", u"32", None))
-                self.STLVoxelization.setEnabled(False)
-                self.LNumberOfVoxelsX.setEnabled(False)
-                self.INNumberOfVoxelsX.setEnabled(False)
-                self.LNumberOfVoxelsY.setEnabled(False)
-                self.INNumberOfVoxelsY.setEnabled(False)
-                self.LNumberOfVoxelsZ.setEnabled(False)
-                self.INNumberOfVoxelsZ.setEnabled(False)
-                self.BVoxelizeGeometry.setEnabled(False)
-            else:
-                warning_message('ERROR: Incorrect file type')
-            pvsimple.Show(self.stl, render_view)
-            pvsimple.ResetCamera(view=None)
-        self.BUploadGeometryFile.clicked.connect(geometry_upload_click)
-        
-        # Voxelize Geometry
-        def voxelize_geometry_click():
-            if not self.INNumberOfVoxelsX.text() or not self.INNumberOfVoxelsY.text() or not self.INNumberOfVoxelsZ.text():
-                warning_message('ERROR: Number of voxels NOT defined')
-            else:
-                try:
-                    self.voxel_reader
-                except:
-                    print('')
-                else:
-                    pvsimple.Delete(self.threshold)
-                    
-                fierro_voxelizer.create_voxel_vtk(
-                    b3_filename[0],
-                    VTK_OUTPUT,
-                    int(self.INNumberOfVoxelsX.text()),
-                    int(self.INNumberOfVoxelsY.text()),
-                    int(self.INNumberOfVoxelsZ.text()),
-                    True
-                )
-                # Paraview window
-                pvsimple.Delete(self.stl)
-                self.voxel_reader = pvsimple.LegacyVTKReader(FileNames = VTK_OUTPUT)
-                pvsimple.SetDisplayProperties(Representation = "Surface")
-                self.threshold = pvsimple.Threshold(Input = self.voxel_reader, Scalars = "density", ThresholdMethod = "Above Upper Threshold", UpperThreshold = 1, LowerThreshold = 0, AllScalars = 1, UseContinuousCellRange = 0, Invert = 0)
-                pvsimple.Show(self.threshold, render_view)
-                pvsimple.Hide(self.voxel_reader)
-                render_view.ResetCamera()
-                render_view.StillRender()
-        self.BVoxelizeGeometry.clicked.connect(voxelize_geometry_click)
-        
-        # Apply Material
-        def material_type():
-            if str(self.INMaterialType.currentText()) == 'Isotropic':
-                self.MaterialTypeTool.setCurrentIndex(0)
-            if str(self.INMaterialType.currentText()) == 'Transversely Isotropic':
-                self.MaterialTypeTool.setCurrentIndex(1)
-            if str(self.INMaterialType.currentText()) == 'Orthotropic':
-                self.MaterialTypeTool.setCurrentIndex(3)
-            if str(self.INMaterialType.currentText()) == 'Anisotropic':
-                self.MaterialTypeTool.setCurrentIndex(2)
-        self.INMaterialType.currentIndexChanged.connect(material_type)
-        
-        def add_material():
-            warning_flag = 0
-            if str(self.INMaterialType.currentText()) == 'Isotropic':
-                if not self.INYoungsModulus.text() or not self.INPoissonsRatio.text() or not self.INMaterialName.text():
-                    warning_message('ERROR: Material definition incomplete')
-                    warning_flag = 1
-                else:
-                    # Calculate Stiffness Matrix
-                    E = float(self.INYoungsModulus.text())
-                    NU = float(self.INPoissonsRatio.text())
-                    INCalcG = float(self.INYoungsModulus.text())/(2*(1+float(self.INPoissonsRatio.text())))
-                    S11 = 1/E
-                    S12 = -NU/E
-                    S13 = -NU/E
-                    S22 = 1/E
-                    S23 = -NU/E
-                    S33 = 1/E
-                    S44 = 1/INCalcG
-                    S55 = 1/INCalcG
-                    S66 = 1/INCalcG
-                    S = S11*S22*S33 - S11*S23*S23 - S22*S13*S13 - S33*S12*S12 + 2*S12*S23*S13
-                    C11 = (1/S)*(S22*S33-S23*S23)
-                    C12 = (1/S)*(S13*S23-S12*S33)
-                    C13 = (1/S)*(S12*S23-S13*S22)
-                    C22 = (1/S)*(S33*S11-S13*S13)
-                    C23 = (1/S)*(S12*S13-S23*S11)
-                    C33 = (1/S)*(S11*S22-S12*S12)
-                    C44 = (1/S44)
-                    C55 = (1/S55)
-                    C66 = (1/S66)
-                
-                    self.INYoungsModulus.clear()
-                    self.INPoissonsRatio.clear()
-            if str(self.INMaterialType.currentText()) == 'Transversely Isotropic':
-                if not self.INEip.text() or not self.INNUip.text() or not self.INEop.text() or not self.INNUop.text() or not self.INGop.text() or not self.INMaterialName.text():
-                    warning_message('ERROR: Material definition incomplete')
-                    warning_flag = 1
-                else:
-                    if str(self.INIsotropicPlane.currentText()) == 'x-y plane':
-                        # Calculate Stiffness Matrix
-                        NUip = float(self.INNUip.text())
-                        NUop = float(self.INNUop.text())
-                        Eip = float(self.INEip.text())
-                        Eop = float(self.INEop.text())
-                        Gop = float(self.INGop.text())
-                        S11 = 1/Eip
-                        S12 = -NUip/Eip
-                        S13 = -NUop/Eop
-                        S22 = 1/Eip
-                        S23 = -NUop/Eop
-                        S33 = 1/Eop
-                        S44 = 1/Gop
-                        S55 = 1/Gop
-                        S66 = 1/(Eip/(2*(1+NUip)))
-                        S = S11*S22*S33 - S11*S23*S23 - S22*S13*S13 - S33*S12*S12 + 2*S12*S23*S13
-                        C11 = (1/S)*(S22*S33-S23*S23)
-                        C12 = (1/S)*(S13*S23-S12*S33)
-                        C22 = (1/S)*(S33*S11-S13*S13)
-                        C13 = (1/S)*(S12*S23-S13*S22)
-                        C33 = (1/S)*(S11*S22-S12*S12)
-                        C23 = (1/S)*(S12*S13-S23*S11)
-                        C44 = (1/S44)
-                        C55 = (1/S55)
-                        C66 = (1/S66)
-                        
-                        self.INEip.clear()
-                        self.INNUip.clear()
-                        self.INEop.clear()
-                        self.INNUop.clear()
-                        self.INGop.clear()
-                    if str(self.INIsotropicPlane.currentText()) == 'x-z plane':
-                        # Calculate Stiffness Matrix
-                        NUip = float(self.INNUip.text())
-                        NUop = float(self.INNUop.text())
-                        Eip = float(self.INEip.text())
-                        Eop = float(self.INEop.text())
-                        Gop = float(self.INGop.text())
-                        S11 = 1/Eip
-                        S12 = -NUop/Eop
-                        S13 = -NUip/Eip
-                        S22 = 1/Eop
-                        S23 = -NUop/Eop
-                        S33 = 1/Eip
-                        S44 = 1/Gop
-                        S55 = 1/(Eip/(2*(1+NUip)))
-                        S66 = 1/Gop
-                        S = S11*S22*S33 - S11*S23*S23 - S22*S13*S13 - S33*S12*S12 + 2*S12*S23*S13
-                        C11 = (1/S)*(S22*S33-S23*S23)
-                        C12 = (1/S)*(S13*S23-S12*S33)
-                        C22 = (1/S)*(S33*S11-S13*S13)
-                        C13 = (1/S)*(S12*S23-S13*S22)
-                        C33 = (1/S)*(S11*S22-S12*S12)
-                        C23 = (1/S)*(S12*S13-S23*S11)
-                        C44 = (1/S44)
-                        C55 = (1/S55)
-                        C66 = (1/S66)
-                        
-                        self.INEip.clear()
-                        self.INNUip.clear()
-                        self.INEop.clear()
-                        self.INNUop.clear()
-                        self.INGop.clear()
-                    if str(self.INIsotropicPlane.currentText()) == 'y-z plane':
-                        # Calculate Stiffness Matrix
-                        NUip = float(self.INNUip.text())
-                        NUop = float(self.INNUop.text())
-                        Eip = float(self.INEip.text())
-                        Eop = float(self.INEop.text())
-                        Gop = float(self.INGop.text())
-                        S11 = 1/Eop
-                        S12 = -NUop/Eop
-                        S13 = -NUop/Eop
-                        S22 = 1/Eip
-                        S23 = -NUip/Eip
-                        S33 = 1/Eip
-                        S44 = 1/(Eip/(2*(1+NUip)))
-                        S55 = 1/Gop
-                        S66 = 1/Gop
-                        S = S11*S22*S33 - S11*S23*S23 - S22*S13*S13 - S33*S12*S12 + 2*S12*S23*S13
-                        C11 = (1/S)*(S22*S33-S23*S23)
-                        C12 = (1/S)*(S13*S23-S12*S33)
-                        C22 = (1/S)*(S33*S11-S13*S13)
-                        C13 = (1/S)*(S12*S23-S13*S22)
-                        C33 = (1/S)*(S11*S22-S12*S12)
-                        C23 = (1/S)*(S12*S13-S23*S11)
-                        C44 = (1/S44)
-                        C55 = (1/S55)
-                        C66 = (1/S66)
-                        
-                        self.INEip.clear()
-                        self.INNUip.clear()
-                        self.INEop.clear()
-                        self.INNUop.clear()
-                        self.INGop.clear()
-    
-            if str(self.INMaterialType.currentText()) == 'Orthotropic':
-                if not self.INEx.text() or not self.INEy.text() or not self.INEz.text() or not self.INNUxy.text() or not self.INNUxz.text() or not self.INNUyz.text() or not self.INGxy.text() or not self.INGxz.text() or not self.INGyz.text() or not self.INMaterialName.text():
-                    warning_message('ERROR: Material definition incomplete')
-                    warning_flag = 1
-                else:
-                    # Calculate Stiffness Matrix
-                    S11 = 1/float(self.INEx.text())
-                    S12 = -float(self.INNUxy.text())/float(self.INEx.text())
-                    S13 = -float(self.INNUxz.text())/float(self.INEx.text())
-                    S22 = 1/float(self.INEy.text())
-                    S23 = -float(self.INNUyz.text())/float(self.INEy.text())
-                    S33 = 1/float(self.INEz.text())
-                    S44 = 1/float(self.INGyz.text())
-                    S55 = 1/float(self.INGxz.text())
-                    S66 = 1/float(self.INGxy.text())
-                    S = S11*S22*S33 - S11*S23*S23 - S22*S13*S13 - S33*S12*S12 + 2*S12*S23*S13
-                    C11 = (1/S)*(S22*S33-S23*S23)
-                    C12 = (1/S)*(S13*S23-S12*S33)
-                    C22 = (1/S)*(S33*S11-S13*S13)
-                    C13 = (1/S)*(S12*S23-S13*S22)
-                    C33 = (1/S)*(S11*S22-S12*S12)
-                    C23 = (1/S)*(S12*S13-S23*S11)
-                    C44 = (1/S44)
-                    C55 = (1/S55)
-                    C66 = (1/S66)
-                    
-                    self.INEx.clear()
-                    self.INEy.clear()
-                    self.INEz.clear()
-                    self.INNUxy.clear()
-                    self.INNUxz.clear()
-                    self.INNUyz.clear()
-                    self.INGxy.clear()
-                    self.INGxz.clear()
-                    self.INGyz.clear()
-                    
-            if str(self.INMaterialType.currentText()) == 'Anisotropic':
-                if not self.TAnisotropic.item(0,0).text() or not self.TAnisotropic.item(0,1).text() or not self.TAnisotropic.item(0,2).text() or not self.TAnisotropic.item(0,3).text() or not self.TAnisotropic.item(0,4).text() or not self.TAnisotropic.item(0,5).text() or not self.TAnisotropic.item(1,1).text() or not self.TAnisotropic.item(1,2).text() or not self.TAnisotropic.item(1,3).text() or not self.TAnisotropic.item(1,4).text() or not self.TAnisotropic.item(1,5).text() or not self.TAnisotropic.item(2,2).text() or not self.TAnisotropic.item(2,3).text() or not self.TAnisotropic.item(2,4).text() or not self.TAnisotropic.item(2,5).text() or not self.TAnisotropic.item(3,3).text() or not self.TAnisotropic.item(3,4).text() or not self.TAnisotropic.item(3,5).text() or not self.TAnisotropic.item(4,4).text() or not self.TAnisotropic.item(4,5).text() or not self.TAnisotropic.item(5,5).text() or not self.INMaterialName.text():
-                    warning_message('ERROR: Material definition incomplete')
-                    warning_flag = 1
-                else:
-                    # Fill out material definition table
-                    row = self.TMaterials.rowCount()
-                    self.TMaterials.insertRow(row)
-                    self.TMaterials.setItem(row, 0, QTableWidgetItem(
-                        self.INMaterialName.text().strip())
-                    )
-                    self.TMaterials.setItem(row, 1, QTableWidgetItem(
-                        str(self.INMaterialType.currentText()))
-                    )
-                    k = 2
-                    for i in [0,1,2,3,4,5,6]:
-                        for j in range(i,6):
-                            self.TMaterials.setItem(
-                                row, k, QTableWidgetItem(self.TAnisotropic.item(i,j).text())
-                            )
-                            self.TAnisotropic.item(i,j).setText('')
-                            k += 1
-                    self.INMaterialName.clear()
-                    warning_flag = 1
-            if warning_flag == 0:
-                # Fill out material definition table
-                row = self.TMaterials.rowCount()
-                self.TMaterials.insertRow(row)
-                self.TMaterials.setItem(row, 0, QTableWidgetItem(
-                    self.INMaterialName.text().strip())
-                )
-                if str(self.INMaterialType.currentText()) == 'Transversely Isotropic':
-                    self.TMaterials.setItem(row, 1, QTableWidgetItem(
-                        str(self.INMaterialType.currentText() + ' ' + self.INIsotropicPlane.currentText()))
-                    )
-                else:
-                    self.TMaterials.setItem(row, 1, QTableWidgetItem(
-                        str(self.INMaterialType.currentText()))
-                    )
-                self.TMaterials.setItem(
-                    row, 2, QTableWidgetItem(str(C11))
-                )
-                self.TMaterials.setItem(
-                    row, 3, QTableWidgetItem(str(C12))
-                )
-                self.TMaterials.setItem(
-                    row, 4, QTableWidgetItem(str(C13))
-                )
-                self.TMaterials.setItem(
-                    row, 8, QTableWidgetItem(str(C22))
-                )
-                self.TMaterials.setItem(
-                    row, 9, QTableWidgetItem(str(C23))
-                )
-                self.TMaterials.setItem(
-                    row, 13, QTableWidgetItem(str(C33))
-                )
-                self.TMaterials.setItem(
-                    row, 17, QTableWidgetItem(str(C44))
-                )
-                self.TMaterials.setItem(
-                    row, 20, QTableWidgetItem(str(C55))
-                )
-                self.TMaterials.setItem(
-                    row, 22, QTableWidgetItem(str(C66))
-                )
-                for i in [5,6,7,10,11,12,14,15,16,18,19,21]:
-                   self.TMaterials.setItem(row, i, QTableWidgetItem('0'))
-                self.INMaterialName.clear()
-            else:
-                warning_flag = 0
-                
-        def delete_material():
-            current_row = self.TMaterials.currentRow()
-            if current_row < 0:
-                return QMessageBox.warning(QMessageBox(),"Warning","Please select a record to delete")
-
-            button = QMessageBox.question(
-                QMessageBox(),
-                'Confirmation',
-                'Are you sure that you want to delete the selected row?',
-                QMessageBox.Yes |
-                QMessageBox.No
-            )
-            if button == QMessageBox.StandardButton.Yes:
-                self.TMaterials.removeRow(current_row)
-                
-        def regenerate_elastic_constants():
-            current_row = self.TMaterials.currentRow()
-            if current_row < 0:
-                return QMessageBox.warning(QMessageBox(),"Warning","Please select a material from the table")
-                
-            # Define Stiffness Matrix
-            Mstiffness = [[float(self.TMaterials.item(current_row,2).text()), float(self.TMaterials.item(current_row,3).text()), float(self.TMaterials.item(current_row,4).text()),  float(self.TMaterials.item(current_row,5).text()), float(self.TMaterials.item(current_row,6).text()), float(self.TMaterials.item(current_row,7).text())], [float(self.TMaterials.item(current_row,3).text()), float(self.TMaterials.item(current_row,8).text()), float(self.TMaterials.item(current_row,9).text()),  float(self.TMaterials.item(current_row,10).text()), float(self.TMaterials.item(current_row,11).text()), float(self.TMaterials.item(current_row,12).text())], [float(self.TMaterials.item(current_row,4).text()), float(self.TMaterials.item(current_row,9).text()), float(self.TMaterials.item(current_row,13).text()), float(self.TMaterials.item(current_row,14).text()), float(self.TMaterials.item(current_row,15).text()), float(self.TMaterials.item(current_row,16).text())], [float(self.TMaterials.item(current_row,5).text()), float(self.TMaterials.item(current_row,10).text()), float(self.TMaterials.item(current_row,14).text()), float(self.TMaterials.item(current_row,17).text()), float(self.TMaterials.item(current_row,18).text()), float(self.TMaterials.item(current_row,19).text())], [float(self.TMaterials.item(current_row,6).text()), float(self.TMaterials.item(current_row,11).text()), float(self.TMaterials.item(current_row,15).text()), float(self.TMaterials.item(current_row,18).text()), float(self.TMaterials.item(current_row,20).text()), float(self.TMaterials.item(current_row,21).text())], [float(self.TMaterials.item(current_row,7).text()), float(self.TMaterials.item(current_row,12).text()), float(self.TMaterials.item(current_row,16).text()), float(self.TMaterials.item(current_row,19).text()), float(self.TMaterials.item(current_row,21).text()), float(self.TMaterials.item(current_row,22).text())]]
-            Mcompliance = np.linalg.inv(Mstiffness)
-            if self.TMaterials.item(current_row,1).text() == 'Isotropic':
-                self.MaterialTypeTool.setCurrentIndex(0)
-                self.INMaterialType.setCurrentIndex(0)
-                self.INMaterialName.clear()
-                self.INYoungsModulus.clear()
-                self.INPoissonsRatio.clear()
-                E = 1/Mcompliance[0][0]
-                nu = -Mcompliance[0][1]*E
-                self.INMaterialName.insert(self.TMaterials.item(current_row,0).text())
-                self.INYoungsModulus.insert(str(E))
-                self.INPoissonsRatio.insert(str(nu))
-            elif 'Transversely Isotropic' in self.TMaterials.item(current_row,1).text():
-                self.MaterialTypeTool.setCurrentIndex(1)
-                self.INMaterialType.setCurrentIndex(1)
-                self.INMaterialName.clear()
-                self.INEip.clear()
-                self.INNUip.clear()
-                self.INEop.clear()
-                self.INNUop.clear()
-                self.INGop.clear()
-                if 'x-y plane' in self.TMaterials.item(current_row,1).text():
-                    Eip = 1/Mcompliance[0][0]
-                    nuip = -Mcompliance[0][1]*Eip
-                    Eop = 1/Mcompliance[2][2]
-                    nuop = -Mcompliance[0][2]*Eop
-                    Gop  = 1/Mcompliance[3][3]
-                    self.INMaterialName.insert(self.TMaterials.item(current_row,0).text())
-                    self.INEip.insert(str(Eip))
-                    self.INNUip.insert(str(nuip))
-                    self.INEop.insert(str(Eop))
-                    self.INNUop.insert(str(nuop))
-                    self.INGop.insert(str(Gop))
-                    self.INIsotropicPlane.setCurrentIndex(0)
-                elif 'x-z plane' in self.TMaterials.item(current_row,1).text():
-                    Eip = 1/Mcompliance[0][0]
-                    nuip = -Mcompliance[0][2]*Eip
-                    Eop = 1/Mcompliance[1][1]
-                    nuop = -Mcompliance[0][1]*Eop
-                    Gop  = 1/Mcompliance[3][3]
-                    self.INMaterialName.insert(self.TMaterials.item(current_row,0).text())
-                    self.INEip.insert(str(Eip))
-                    self.INNUip.insert(str(nuip))
-                    self.INEop.insert(str(Eop))
-                    self.INNUop.insert(str(nuop))
-                    self.INGop.insert(str(Gop))
-                    self.INIsotropicPlane.setCurrentIndex(1)
-                elif 'y-z plane' in self.TMaterials.item(current_row,1).text():
-                    Eip = 1/Mcompliance[1][1]
-                    nuip = -Mcompliance[1][2]*Eip
-                    Eop = 1/Mcompliance[0][0]
-                    nuop = -Mcompliance[0][1]*Eop
-                    Gop  = 1/Mcompliance[4][4]
-                    self.INMaterialName.insert(self.TMaterials.item(current_row,0).text())
-                    self.INEip.insert(str(Eip))
-                    self.INNUip.insert(str(nuip))
-                    self.INEop.insert(str(Eop))
-                    self.INNUop.insert(str(nuop))
-                    self.INGop.insert(str(Gop))
-                    self.INIsotropicPlane.setCurrentIndex(2)
-            elif self.TMaterials.item(current_row,1).text() == 'Orthotropic':
-                self.MaterialTypeTool.setCurrentIndex(3)
-                self.INMaterialType.setCurrentIndex(2)
-                self.INMaterialName.clear()
-                self.INEx.clear()
-                self.INEy.clear()
-                self.INEz.clear()
-                self.INNUxy.clear()
-                self.INNUxz.clear()
-                self.INNUyz.clear()
-                self.INGxy.clear()
-                self.INGxz.clear()
-                self.INGyz.clear()
-                Ex = 1/Mcompliance[0][0]
-                Ey = 1/Mcompliance[1][1]
-                Ez = 1/Mcompliance[2][2]
-                NUxy = -Mcompliance[0][1]*Ex
-                NUxz = -Mcompliance[0][2]*Ex
-                NUyz = -Mcompliance[1][2]*Ey
-                Gxy = 1/Mcompliance[5][5]
-                Gxz = 1/Mcompliance[4][4]
-                Gyz = 1/Mcompliance[3][3]
-                self.INMaterialName.insert(self.TMaterials.item(current_row,0).text())
-                self.INEx.insert(str(Ex))
-                self.INEy.insert(str(Ey))
-                self.INEz.insert(str(Ez))
-                self.INNUxy.insert(str(NUxy))
-                self.INNUxz.insert(str(NUxz))
-                self.INNUyz.insert(str(NUyz))
-                self.INGxy.insert(str(Gxy))
-                self.INGxz.insert(str(Gxz))
-                self.INGyz.insert(str(Gyz))
-            else:
-                self.MaterialTypeTool.setCurrentIndex(2)
-                self.INMaterialType.setCurrentIndex(3)
-                self.INMaterialName.clear()
-                self.INMaterialName.insert(self.TMaterials.item(current_row,0).text())
-                k = 2
-                for i in [0,1,2,3,4,5,6]:
-                    for j in range(i,6):
-                        self.TAnisotropic.item(i,j).setText('')
-                        self.TAnisotropic.setItem(
-                            i, j, QTableWidgetItem(self.TMaterials.item(current_row,k).text())
-                        )
-                        k += 1
-                        
-        self.BAddMaterial.clicked.connect(add_material)
-        self.BDeleteMaterial.clicked.connect(delete_material)
-        self.BRegenElasticConstants.clicked.connect(regenerate_elastic_constants)
-        
-        # Boundary Conditions
-        def BC_direction():
-            if self.INBoundaryCondition.currentText() == "Tension" or self.INBoundaryCondition.currentText() == "Compression":
-                self.INBCDirection.setItemText(0, QCoreApplication.translate("MainWindow", u"x-direction", None))
-                self.INBCDirection.setItemText(1, QCoreApplication.translate("MainWindow", u"y-direction", None))
-                self.INBCDirection.setItemText(2, QCoreApplication.translate("MainWindow", u"z-direction", None))
-            elif self.INBoundaryCondition.currentText() == "Shear":
-                self.INBCDirection.setItemText(0, QCoreApplication.translate("MainWindow", u"xy-direction", None))
-                self.INBCDirection.setItemText(1, QCoreApplication.translate("MainWindow", u"xz-direction", None))
-                self.INBCDirection.setItemText(2, QCoreApplication.translate("MainWindow", u"yz-direction", None))
-        self.INBoundaryCondition.currentTextChanged.connect(BC_direction)
-        
-        def add_bcs():
-#            if not self.INBCStrainRate.text():
-#                on_clicked("ERROR: Boundary condition incomplete")
-#            else:
-            row = self.TBCs.rowCount()
-            self.TBCs.insertRow(row)
-            self.TBCs.setItem(row, 0, QTableWidgetItem(str(
-                self.INBoundaryCondition.currentText()))
-            )
-            self.TBCs.setItem(
-                row, 1, QTableWidgetItem(str(self.INBCDirection.currentText()))
-            )
-#            self.TBCs.setItem(
-#                row, 2, QTableWidgetItem(self.INBCStrainRate.text())
-#            )
-#            reset_bcs()
-#
-#        def reset_bcs():
-#            self.INBCStrainRate.clear()
-
-        def delete_bcs():
-            current_row = self.TBCs.currentRow()
-            if current_row < 0:
-                return QMessageBox.warning(self, 'Warning','Please select a record to delete')
-
-            button = QMessageBox.question(
-                QMessageBox(),
-                'Confirmation',
-                'Are you sure that you want to delete the selected row?',
-                QMessageBox.StandardButton.Yes |
-                QMessageBox.StandardButton.No
-            )
-            if button == QMessageBox.StandardButton.Yes:
-                self.TBCs.removeRow(current_row)
-                
-        self.BAddBC.clicked.connect(add_bcs)
-        self.BDeleteBC.clicked.connect(delete_bcs)
-        
-        # Write input file
-        def write_input_file():
-            # Get current file location
-            current_file_loc = os.getcwd()
-            
-            # Elastic parameters file
-#            if self.TMaterials.item(0,1).text() == 'Isotropic':
-#                elastic_parameters = open(ELASTIC_PARAMETERS,"w")
-#                iso = '1                         ISO\n'
-#                elastic_parameters.write(iso)
-#                Young_Nu = self.TMaterials.item(0,2).text() + '   ' + self.TMaterials.item(0,5).text() + '                  YOUNG(MPa),NU (V+R/2)\n'
-#                elastic_parameters.write(Young_Nu)
-#                elastic_parameters.close()
-            if self.TMaterials.item(0,1).text() == 'Isotropic' or 'Transversely Isotropic' in self.TMaterials.item(0,1).text() or self.TMaterials.item(0,1).text() == 'Orthotropic':
-                elastic_parameters = open(ELASTIC_PARAMETERS,"w")
-                iso = '0\n'
-                elastic_parameters.write(iso)
-                stiffness = '  ' + self.TMaterials.item(0,2).text() + '  ' + self.TMaterials.item(0,3).text() + '  ' + self.TMaterials.item(0,4).text() + '  0  0  0     Cu (MPa)\n' + '  ' + self.TMaterials.item(0,3).text() + '  ' + self.TMaterials.item(0,8).text() + '  ' + self.TMaterials.item(0,9).text() + '  0  0  0\n' + '  ' + self.TMaterials.item(0,4).text() + '  ' + self.TMaterials.item(0,9).text() + '  ' + self.TMaterials.item(0,13).text() + '  0  0  0\n' + '  0  0  0   ' + self.TMaterials.item(0,17).text() + '  0  0\n' + '  0  0  0  0  ' + self.TMaterials.item(0,20).text() + '  0\n' + '  0  0  0  0  0  ' + self.TMaterials.item(0,22).text()
-                elastic_parameters.write(stiffness)
-                elastic_parameters.close()
-            elif self.TMaterials.item(0,1).text() == 'Anisotropic':
-                elastic_parameters = open(ELASTIC_PARAMETERS,"w")
-                iso = '0\n'
-                elastic_parameters.write(iso)
-                stiffness = '  ' + self.TMaterials.item(0,2).text() + '  ' + self.TMaterials.item(0,3).text() + '  ' + self.TMaterials.item(0,4).text() + '  ' + self.TMaterials.item(0,5).text() + '  ' + self.TMaterials.item(0,6).text() + '  ' + self.TMaterials.item(0,7).text() + '     Cu (MPa)\n' + '  ' + self.TMaterials.item(0,3).text() + '  ' + self.TMaterials.item(0,8).text() + '  ' + self.TMaterials.item(0,9).text() + '  ' + self.TMaterials.item(0,10).text() + '  ' + self.TMaterials.item(0,11).text() + '  ' + self.TMaterials.item(0,12).text() + '\n' + '  ' + self.TMaterials.item(0,4).text() + '  ' + self.TMaterials.item(0,9).text() + '  ' + self.TMaterials.item(0,13).text() + '  ' + self.TMaterials.item(0,14).text() + '  ' + self.TMaterials.item(0,15).text() + '  ' + self.TMaterials.item(0,16).text() + '\n' + '  ' + self.TMaterials.item(0,5).text() + '  ' + self.TMaterials.item(0,10).text() + '  ' + self.TMaterials.item(0,14).text() + '  ' + self.TMaterials.item(0,17).text() + '  ' + self.TMaterials.item(0,18).text() + '  ' + self.TMaterials.item(0,19).text() + '\n' + '  ' + self.TMaterials.item(0,6).text() + '  ' + self.TMaterials.item(0,11).text() + '  ' + self.TMaterials.item(0,15).text() + '  ' + self.TMaterials.item(0,18).text() + '  ' + self.TMaterials.item(0,20).text() + '  ' + self.TMaterials.item(0,21).text() + '\n' + '  ' + self.TMaterials.item(0,7).text() + '  ' + self.TMaterials.item(0,12).text() + '  ' + self.TMaterials.item(0,16).text() + '  ' + self.TMaterials.item(0,19).text() + '  ' + self.TMaterials.item(0,21).text() + '  ' +  self.TMaterials.item(0,22).text()
-                elastic_parameters.write(stiffness)
-                elastic_parameters.close()
-            
-            # EVPFFT lattice input parameters file
-            evpfft_lattice_input = open(EVPFFT_INPUT,"w")
-            modes = '2 0 0 0               NPHMX, NMODMX, NTWMMX, NSYSMX\n'
-            evpfft_lattice_input.write(modes)
-            dimensions = str(int(self.INNumberOfVoxelsX.text())) + ' ' + str(int(self.INNumberOfVoxelsY.text())) + ' ' + str(int(self.INNumberOfVoxelsZ.text())) + '               x-dim, y-dim, z-dim\n'
-            evpfft_lattice_input.write(dimensions)
-            nph_delt = '2                      number of phases (nph)\n' + '1.  1.  1.             RVE dimensions (delt)\n' + '* name and path of microstructure file (filetext)\n'
-            evpfft_lattice_input.write(nph_delt)
-            vtkfile = f'{VTK_OUTPUT}\n'
-            evpfft_lattice_input.write(vtkfile)
-            phases = '*INFORMATION ABOUT PHASE #1\n' + '1                          igas(iph)\n' + '* name and path of single crystal files (filecryspl, filecrysel) (dummy if igas(iph)=1)\n' + 'dummy\n' + 'dummy\n' + '*INFORMATION ABOUT PHASE #2\n' + '0                          igas(iph)\n' + '* name and path of single crystal files (filecryspl, filecrysel) (dummy if igas(iph)=1)\n' +  f'{PLASTIC_PARAMETERS}\n' + f'{ELASTIC_PARAMETERS}\n'
-            evpfft_lattice_input.write(phases)
-            if str(self.INBCDirection.currentText()) == "x-direction":
-                if str(self.INBoundaryCondition.currentText()) == "Tension":
-                    test_conditions = '*INFORMATION ABOUT TEST CONDITIONS\n' + '* boundary conditions\n' + '    1       1       1           iudot     |    flag for vel.grad.\n' + '    1       0       1                     |    (0:unknown-1:known)\n' + '    1       1       0                     |\n' + '                                          |\n' + '   1.0     0.        0.          udot     |    vel.grad\n' + '    0.      0.      0.                  |\n' + '    0.       0.         0.                |\n' + '                                          |\n' + '    0       0        0           iscau    |    flag for Cauchy\n' + '            1        0                    |\n' + '                     1                    |\n' + '                                          |\n' + '    0.      0.       0.          scauchy  |    Cauchy stress\n' + '            0.       0.                   |\n' + '                     0.                   @\n'
-                elif str(self.INBoundaryCondition.currentText()) == "Compression":
-                    test_conditions = '*INFORMATION ABOUT TEST CONDITIONS\n' + '* boundary conditions\n' + '    1       1       1           iudot     |    flag for vel.grad.\n' + '    1       0       1                     |    (0:unknown-1:known)\n' + '    1       1       0                     |\n' + '                                          |\n' + '   -1.0     0.        0.          udot    |    vel.grad\n' + '    0.      0.      0.                  |\n' + '    0.       0.         0.                |\n' + '                                          |\n' + '    0       0        0           iscau    |    flag for Cauchy\n' + '            1        0                    |\n' + '                     1                    |\n' + '                                          |\n' + '    0.      0.       0.          scauchy  |    Cauchy stress\n' + '            0.       0.                   |\n' + '                     0.                   @\n'
-                else:
-                    print("INVALID BOUNDARY CONDITION")
-            elif str(self.INBCDirection.currentText()) == "y-direction":
-                if str(self.INBoundaryCondition.currentText()) == "Tension":
-                    test_conditions = '*INFORMATION ABOUT TEST CONDITIONS\n' + '* boundary conditions\n' + '    0       1       1           iudot     |    flag for vel.grad.\n' + '    1       1       1                     |    (0:unknown-1:known)\n' + '    1       1       0                     |\n' + '                                          |\n' + '   0.     0.        0.          udot    |    vel.grad\n' + '    0.      1.0      0.                  |\n' + '    0.       0.         0.                |\n' + '                                          |\n' + '    1       0        0           iscau    |    flag for Cauchy\n' + '            0        0                    |\n' + '                     1                    |\n' + '                                          |\n' + '    0.      0.       0.          scauchy  |    Cauchy stress\n' + '            0.       0.                   |\n' + '                     0.                   @\n'
-                elif str(self.INBoundaryCondition.currentText()) == "Compression":
-                    test_conditions = '*INFORMATION ABOUT TEST CONDITIONS\n' + '* boundary conditions\n' + '    0       1       1           iudot     |    flag for vel.grad.\n' + '    1       1       1                     |    (0:unknown-1:known)\n' + '    1       1       0                     |\n' + '                                          |\n' + '   0.     0.        0.          udot    |    vel.grad\n' + '    0.      -1.0      0.                  |\n' + '    0.       0.         0.                |\n' + '                                          |\n' + '    1       0        0           iscau    |    flag for Cauchy\n' + '            0        0                    |\n' + '                     1                    |\n' + '                                          |\n' + '    0.      0.       0.          scauchy  |    Cauchy stress\n' + '            0.       0.                   |\n' + '                     0.                   @\n'
-                else:
-                    print("INVALID BOUNDARY CONDITION")
-            elif str(self.INBCDirection.currentText()) == "z-direction":
-                if str(self.INBoundaryCondition.currentText()) == "Tension":
-                    test_conditions = '*INFORMATION ABOUT TEST CONDITIONS\n' + '* boundary conditions\n' + '    0       1       1           iudot     |    flag for vel.grad.\n' + '    1       0       1                     |    (0:unknown-1:known)\n' + '    1       1       1                     |\n' + '                                          |\n' + '   0.     0.        0.          udot    |    vel.grad\n' + '    0.      0.      0.                  |\n' + '    0.       0.         1.0                |\n' + '                                          |\n' + '    1       0        0           iscau    |    flag for Cauchy\n' + '            1        0                    |\n' + '                     0                    |\n' + '                                          |\n' + '    0.      0.       0.          scauchy  |    Cauchy stress\n' + '            0.       0.                   |\n' + '                     0.                   @\n'
-                elif str(self.INBoundaryCondition.currentText()) == "Compression":
-                    test_conditions = '*INFORMATION ABOUT TEST CONDITIONS\n' + '* boundary conditions\n' + '    0       1       1           iudot     |    flag for vel.grad.\n' + '    1       0       1                     |    (0:unknown-1:known)\n' + '    1       1       1                     |\n' + '                                          |\n' + '   0.0     0.        0.          udot    |    vel.grad\n' + '    0.      0.0      0.                  |\n' + '    0.       0.         -1.0                |\n' + '                                          |\n' + '    1       0        0           iscau    |    flag for Cauchy\n' + '            1        0                    |\n' + '                     0                    |\n' + '                                          |\n' + '    0.      0.       0.          scauchy  |    Cauchy stress\n' + '            0.       0.                   |\n' + '                     0.                   @\n'
-                else:
-                    print("INVALID BOUNDARY CONDITION")
-            elif str(self.INBoundaryCondition.currentText()) == "Shear":
-                if str(self.INBCDirection.currentText()) == "xy-direction":
-                    test_conditions = '*INFORMATION ABOUT TEST CONDITIONS\n' + '* boundary conditions\n' + '    1       1       1           iudot     |    flag for vel.grad.\n' + '    1       1       1                     |    (0:unknown-1:known)\n' + '    1       1       1                     |\n' + '                                          |\n' + '   0.     1.0        0.          udot    |    vel.grad\n' + '    1.0      0.      0.                  |\n' + '    0.       0.         0.                |\n' + '                                          |\n' + '    0       0        0           iscau    |    flag for Cauchy\n' + '            0        0                    |\n' + '                     0                    |\n' + '                                          |\n' + '    0.      0.       0.          scauchy  |    Cauchy stress\n' + '            0.       0.                   |\n' + '                     0.                   @\n'
-                elif str(self.INBCDirection.currentText()) == "xz-direction":
-                    test_conditions = '*INFORMATION ABOUT TEST CONDITIONS\n' + '* boundary conditions\n' + '    1       1       1           iudot     |    flag for vel.grad.\n' + '    1       1       1                     |    (0:unknown-1:known)\n' + '    1       1       1                     |\n' + '                                          |\n' + '   0.     0.        1.0          udot    |    vel.grad\n' + '    0.      0.      0.                  |\n' + '    1.0       0.         0.                |\n' + '                                          |\n' + '    0       0        0           iscau    |    flag for Cauchy\n' + '            0        0                    |\n' + '                     0                    |\n' + '                                          |\n' + '    0.      0.       0.          scauchy  |    Cauchy stress\n' + '            0.       0.                   |\n' + '                     0.                   @\n'
-                elif str(self.INBCDirection.currentText()) == "yz-direction":
-                    test_conditions = '*INFORMATION ABOUT TEST CONDITIONS\n' + '* boundary conditions\n' + '    1       1       1           iudot     |    flag for vel.grad.\n' + '    1       1       1                     |    (0:unknown-1:known)\n' + '    1       1       1                     |\n' + '                                          |\n' + '   0.     0.        0.          udot    |    vel.grad\n' + '    0.      0.      1.0                  |\n' + '    0.       1.0         0.                |\n' + '                                          |\n' + '    0       0        0           iscau    |    flag for Cauchy\n' + '            0        0                    |\n' + '                     0                    |\n' + '                                          |\n' + '    0.      0.       0.          scauchy  |    Cauchy stress\n' + '            0.       0.                   |\n' + '                     0.                   @\n'
-            else:
-                orint("INVALID BOUNDARY CONDITION")
-            evpfft_lattice_input.write(test_conditions)
-            other = '* other\n' + '0.0001         eqincr (if ictrl>=0) or tdot (if ictrl=-1)\n' + '-1              ictrl (1-6: strain comp, 0: VM eq, -1: tdot)\n'
-            evpfft_lattice_input.write(other)
-            run_conditions = '*INFORMATION ABOUT RUN CONDITIONS\n' + self.INNumberOfSteps.text() + '              nsteps\n' + '0.00001         err\n' + '50              itmax\n' + '0               IRECOVER read grain states from STRESS.IN  (1) or not (0)?\n' + '0               ISAVE write grain states in STRESS.OUT (1) or not (0)?\n' + '1               IUPDATE update tex & RVE dim (1) or not (0)?\n' + '0               IUPHARD\n' + '1               IWTEX\n' + '1 10            IWFIELDS,IWSTEP\n' + '0               ITHERMO (if ithermo=1, next line is filethermo)\n' + 'dummy\n'
-            evpfft_lattice_input.write(run_conditions)
-            evpfft_lattice_input.close()
-            
-        # Run EVPFFT
-        self.run_cnt = 0
-        def run_click():
-            write_input_file()
-            self.p = QProcess()
-            self.p.readyReadStandardOutput.connect(handle_stdout)
-            self.p.readyReadStandardError.connect(handle_stderr)
-            self.p.stateChanged.connect(handle_state)
-            self.p.finished.connect(process_finished)
-            self.p.start("evpfft",["-f", EVPFFT_INPUT, "-m", "2"])
-            self.progress_re = re.compile("       Current  Time  STEP = (\d+)")
-            self.run_cnt += 1
-        def simple_percent_parser(output):
-            m = self.progress_re.search(output)
-            if m:
-                pc_complete = m.group(1)
-                return int(pc_complete)
-        def process_finished():
-            self.RunOutputProgress.setValue(100)
-            self.p = None
-        def handle_stdout():
-            data = self.p.readAllStandardOutput()
-            stdout = bytes(data).decode("utf8")
-            progress = simple_percent_parser(stdout)
-            if progress:
-                self.RunOutputProgress.setValue((progress/int(self.INNumberOfSteps.text()))*100)
-            self.RunOutputWindow.appendPlainText(stdout)
-        def handle_stderr():
-            data = self.p.readAllStandardError()
-            stderr = bytes(data).decode("utf8")
-            self.RunOutputWindow.appendPlainText(stderr)
-        def handle_state(state):
-            states = {
-                QProcess.NotRunning: 'Finished',
-                QProcess.Starting: 'Starting EVPFFT',
-                QProcess.Running: 'Running EVPFFT',
-            }
-            state_name = states[state]
-            self.RunOutputWindow.appendPlainText(f"{state_name}")
-            
-        self.BRunEVPFFT.clicked.connect(run_click)
-        
-        
-        # Preview Results
-        def preview_results_click():
-            # Delete previous views
-            try:
-                self.stl
-            except:
-                print('')
-            else:
-                pvsimple.Delete(self.stl)
-            try:
-                self.threshold
-            except:
-                print('')
-            else:
-                pvsimple.Delete(self.threshold)
-            try:
-                self.threshold2
-            except:
-                print('')
-            else:
-                pvsimple.Delete(self.threshold2)
-            # Render new view
-            self.results_reader = pvsimple.XDMFReader(FileNames = "micro_state_timestep_10.xdmf")
-            pvsimple.SetDisplayProperties(Representation = "Surface")
-            self.threshold2 = pvsimple.Threshold(registrationName='results_threshold', Input = self.results_reader, Scalars = "phase_id", ThresholdMethod = "Above Upper Threshold", UpperThreshold = 2, LowerThreshold = 1, AllScalars = 1, UseContinuousCellRange = 0, Invert = 0)
-            display = pvsimple.Show(self.threshold2, render_view)
-            # Select what variable you want to display
-            pvsimple.GetAnimationScene().GoToLast()
-            pvsimple.ColorBy(display,('CELLS',str(self.INPreviewResults.currentText())))
-            vmstressLUT = pvsimple.GetColorTransferFunction(str(self.INPreviewResults.currentText()))
-            r = self.results_reader.CellData.GetArray(str(self.INPreviewResults.currentText())).GetRange()
-            vmstressLUT.RescaleTransferFunction(r[0], r[1]/2)
-            display.SetScalarBarVisibility(render_view, True)
-            pvsimple.HideUnusedScalarBars(render_view)
-            # Add time filter
-            threshold1 = pvsimple.FindSource('results_threshold')
-            annotateTimeFilter1 = pvsimple.AnnotateTimeFilter(registrationName='AnnotateTimeFilter1', Input=threshold1)
-            annotateTimeFilter1Display = pvsimple.Show(annotateTimeFilter1, render_view, 'TextSourceRepresentation')
-            # Remove old view / reset cameras
-            pvsimple.Hide(self.results_reader)
-            render_view.ResetCamera()
-            render_view.StillRender()
-        self.BPreviewResults.clicked.connect(preview_results_click)
-        self.BPreviewResults.clicked.connect(lambda: self.OutputWindows.setCurrentIndex(0))
-        
-        # Stress vs Strain Plot
-        def plot_ss_click():
-            # Get the stress-strain data
-            try:
-                self.Plot.figure
-            except:
-                with open("str_str.out", newline='') as f:
-                    reader = csv.reader(f)
-                    self.ss_data = list(reader)
-                x = [0 for i in range(int(self.INNumberOfSteps.text()))]
-                y = [0 for i in range(int(self.INNumberOfSteps.text()))]
-                for i in range(int(self.INNumberOfSteps.text())):
-                    if str(self.INPlotSS.currentText()) == 'S11 vs E11':
-                        xcol = 0
-                        ycol = 6
-                    elif str(self.INPlotSS.currentText()) == 'S22 vs E22':
-                        xcol = 1
-                        ycol = 7
-                    elif str(self.INPlotSS.currentText()) == 'S33 vs E33':
-                        xcol = 2
-                        ycol = 8
-                    x[i] = float(self.ss_data[i+1][xcol])
-                    y[i] = float(self.ss_data[i+1][ycol])
-                # Plot data
-                self.Plot.figure = Figure()
-                self.Plot.ax = self.Plot.figure.add_subplot()
-                self.Plot.ax.plot(x,y)
-                self.Plot.ax.set_xlabel('STRAIN')
-                self.Plot.ax.set_ylabel('STRESS')
-                self.Plot.figure.tight_layout()
-                # Display plot and toolbar
-                layout = QVBoxLayout()
-                self.Plot.setLayout(layout)
-                self.Plot.canvas = FigureCanvasQTAgg(self.Plot.figure)
-                layout.addWidget(self.Plot.canvas)
-                self.toolbar = NavigationToolbar2QT(self.Plot.canvas,self.Plot)
-                layout.addWidget(self.toolbar)
-            else:
-                self.timer = QTimer()
-                self.timer.setInterval(100)
-                self.timer.timeout.connect(update_plot)
-                self.timer.start()
-        def update_plot():
-            if self.run_cnt > 1:
-                with open("str_str.out", newline='') as f:
-                    reader = csv.reader(f)
-                    self.ss_data = list(reader)
-            x = [0 for i in range(int(self.INNumberOfSteps.text()))]
-            y = [0 for i in range(int(self.INNumberOfSteps.text()))]
-            for i in range(int(self.INNumberOfSteps.text())):
-                if str(self.INPlotSS.currentText()) == 'S11 vs E11':
-                    xcol = 0
-                    ycol = 6
-                elif str(self.INPlotSS.currentText()) == 'S22 vs E22':
-                    xcol = 1
-                    ycol = 7
-                elif str(self.INPlotSS.currentText()) == 'S33 vs E33':
-                    xcol = 2
-                    ycol = 8
-                x[i] = float(self.ss_data[i+1][xcol])
-                y[i] = float(self.ss_data[i+1][ycol])
-            self.Plot.ax.cla()
-            self.Plot.ax.plot(x,y)
-            self.Plot.ax.set_xlabel('STRAIN')
-            self.Plot.ax.set_ylabel('STRESS')
-            self.Plot.figure.tight_layout()
-            self.Plot.canvas.draw()
-            self.timer.stop()
-        self.BPlotSS.clicked.connect(plot_ss_click)
-        self.BPlotSS.clicked.connect(lambda: self.OutputWindows.setCurrentIndex(1))
-        
-        # Open Paraview
-        def open_paraview_click():
-            os.system("open " + "micro_state_timestep_10.xdmf")
-        self.BOpenParaview.clicked.connect(open_paraview_click)
-        
-        # Warning Message Popup
-        def warning_message(msg):
-            message = QMessageBox()
-            message.setText(msg)
-            message.exec()
-        
-    # =================================================================
         
     # setupUi
